# Speed up dataset tests, compressed writes to zarr3 arrays