--- a/webknossos/pyproject.toml
+++ b/webknossos/pyproject.toml
@@ -64,6 +64,7 @@ JPype1 = { version = "^1.3.0", optional = true }
 pims = { version = "^0.6.0", optional = true }
 tifffile = { version = ">=2021.11.2", optional = true }
 pylibCZIrw = { version = "3.5.1", source = "scm", optional = true }
+pytest-timestamper = "^0.0.9"
 
 [tool.poetry.extras]
 pims = ["pims"]